sdk/pyproject.toml
[build-system]
requires = ["setuptools>=78.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "movieIsdk"
version = "0.0.2"
authors = [
  { name="[Isabelle LACMAGO]" },
]
description = "SDK Python pour accéder à l’API MovieLens (Data Analysts/Data Scientists friendly)"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
       'httpx>=0.28.1',
       'pydantic>=2.11.4',
       'backoff>=2.2.1',
       'numpy>=2.2.4',
       'pandas>=2.2.3',
       'python-dotenv',
]